+++ docs/download_resources.jl
@@ -35,6 +35,7 @@ fetch_assets(
                 "heat_equation", "linear_elasticity", "incompressible_elasticity",
                 "incompressible_elasticity_pressure",
                 "hyperelasticity", "plasticity", "computational_homogenization",
+                "stress_driven_homogenization",
                 "stokes-flow", "dg_heat_equation", "linear_shell",
                 "linear_elasticity_stress", "elastodynamics_modes", "darcy_flow",
             ], ".png"