+++ src/Ferrite.jl
@@ -129,6 +129,7 @@ include("Grid/grid_generators.jl")
 include("Grid/coloring.jl")
 
 # Dofs
+include("Dofs/algebraic_variables.jl")
 include("Dofs/DofHandler.jl")
 include("Dofs/ConstraintHandler.jl")
 include("Dofs/apply_analytical.jl")
@@ -137,6 +138,7 @@ include("Dofs/block_sparsity_pattern.jl")
 include("Dofs/DofRenumbering.jl")
 
 include("iterators.jl")
+include("Dofs/algebraic_coupling.jl")
 
 # Assembly
 include("assembler.jl")